requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.28.0
flask>=3.0.0
gradio>=4.0.0
grad-cam>=1.4.8
tqdm>=4.65.0
pandas>=2.0.0
seaborn>=0.12.0
ultralytics>=8.0.0
